confirm message for deleting expenses

+++ src/app/dashboard/dashboard.component.ts
@@ -46,15 +46,17 @@ export class DashboardComponent {
     }
 
     deleteExpense(id: string) {
-        this.http.delete(`https://localhost:7265/api/Expenses/${id}`)
-            .subscribe({
-                // Will only run when we get a success response from API
-                next: (response) => {
-                    this.expenseList$ = this.getExpenses();
-                    // alert("Expense deleted successfully.");
-                    this.ngOnInit();
-                }
-            });
+        if (window.confirm("Are you sure you want to delete this expense?")) {  //Confirmation request
+            this.http.delete(`https://localhost:7265/api/Expenses/${id}`)
+                .subscribe({
+                    // Will only run when we get a success response from API
+                    next: (response) => {
+                        this.expenseList$ = this.getExpenses();
+                        // alert("Expense deleted successfully.");
+                        this.ngOnInit();
+                    }
+                });
+        }
     }
 
     private getExpenses(): Observable<Expense[]> {
